requirements_new.txt
faiss-cpu
openai
numpy
pandas
sentence-transformers
python-dotenv
gradio